pyproject.toml
[project]
name = "deepseek-hf-monitor"
version = "0.1.0"
description = "Monitor new models under a Hugging Face org (default: deepseek-ai)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
]

[project.scripts]
deepseek-monitor = "deepseek_hf_monitor.monitor:main"
deepseek-monitor-healthcheck = "deepseek_hf_monitor.healthcheck:main"

[dependency-groups]
dev = [
  "pytest>=8.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deepseek_hf_monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
